Restructure namespaces in AccountManagement.Tests project to reflect the new vertical sliced architecture
